Revert "Lots of moving around... almost working. Try src directory again, same format as root CMakeLists.txt inside src, basic one outside."
This reverts commit 2035aba3bbcd3448b02696bb70bb9aaa58e17eac.

+++ src/Types.h
@@ -102,7 +102,7 @@ inline EPiece getPiece(PieceType p, EColor color) {
 //}
 
 inline int pop_LSB(int64 &b) {
-    int i = (int) get_LSB(b);
+    int i = get_LSB(b);
     b &= b - 1;
     return i;
 }
